The output json file is in the right format now
The with open('airfare#1.json', 'w+') doesn't read the json data properly when it should.
Therefore I have added: with open('airfare#1.json', 'r') above it to open get the json data properly.
After that, the file is going to be opened like this: with open('airfare#1.json', 'w') and add the new json data taken from the API.