This file unites both the functionality of the 'data_getter' and the 'data_analyser'.
I did it like this in order to make the functionality more centralised and this is arguably the most abstracted way to actually way of getting to the end point. The end point is where you have you have your json files filled with data and a matplotlib graph is plotted based on that data.